refactor: split sensor descriptors and setup
